requirements.txt
absl-py==0.10.0
astunparse==1.6.3
atari-py==0.2.6
box2d-py==2.3.8
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.1
cycler==0.10.0
Cython==0.29.21
docker-pycreds==0.4.0
fasteners==0.15
filelock==3.0.12
future==0.18.2
gast==0.4.0
gitdb==4.0.5
GitPython==3.1.9
glfw==2.0.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
h5py==2.10.0
idna==2.10
imageio==2.9.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.3.1
matplotlib==3.3.2
monotonic==1.5
mujoco-py==2.0.2.13
nltk==3.5
numpy==1.19.2
numpy-ringbuffer==0.2.1
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.3
pathtools==0.1.2
Pillow==7.2.0
promise==2.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.11
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scipy==1.5.2
sentry-sdk==0.19.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
stable-baselines3==0.11.1
subprocess32==3.5.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tokenizers==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
transformers==4.3.3
urllib3==1.25.10
wandb==0.10.5
watchdog==0.10.3
Werkzeug==1.0.1
wrapt==1.12.1
